remove hair component and simplify model component
With the way that the hair support is currently woven through the
renderer, serializing, model and mesh code it's making refactoring
things quite complex. We know the code is preserved in Git so for now it
helps simplify things to remove and hopefully we can find a way of
bringing the code back in some more orthogonal way later.